Move to multiple JSON files
Move to use a JSON file in each folder that has content as opposed to
a single JSON at the top level and remove the -j option from the argparse.
Use a file glob to seach for all "wrapper_input.json" files found below the
source directory.

Change all results objects to lists
Add a "-n" option to print out the names of the files that will be created and
, at this point, do not continue with generation.