[FEM] FemSolver and static elasticity test (#14649)
* Add FemSolver and static elasticity test

FemSolver provides the minimum functionalities needed to solve the
system of equations that arise from FEM space discretization by
integrating components such as FemModel, LinearSystemSolver and
DirichletBoundaryCondition.

FemSolverTest serves as an integration test for FEM. It generates
an arbitrary set of forces on the vertices of a model by setting
arbitrary displacements for the vertices of the model. It then verifies
that under those forces, the solution provided the solver matches the
prescribed displacements.